move P-Roc devices in their own file. add docstrings
